Migrate v2 'zk-read' APIs to 'api' module
