+++ solr/api/src/java/org/apache/solr/client/api/model/ZooKeeperFileResponse.java
@@ -1,0 +1,28 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.solr.client.api.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class ZooKeeperFileResponse extends SolrJerseyResponse {
+  // TODO Should be switched over to using StreamingOutput as a part of SOLR-17562
+  @JsonProperty("content") // A flag value that RawResponseWriter handles specially
+  public Object output;
+
+  @JsonProperty("zkData")
+  public String zkData;
+}
