+++ solr/api/src/java/org/apache/solr/client/api/model/ZooKeeperListChildrenResponse.java
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.solr.client.api.model;
+
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
+import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.HashMap;
+import java.util.Map;
+
+public class ZooKeeperListChildrenResponse extends ExperimentalResponse {
+  @JsonProperty("stat")
+  public ZooKeeperStat stat;
+
+  // TODO Currently the list response (when child information is fetched) consists primarily of an
+  //  object with only one key - the name of the root node - with separate objects under there for
+  //  each child.  The additional nesting under the root node doesn't serve much purpose afaict
+  //  and should be removed.
+  public Map<String, Map<String, ZooKeeperStat>> unknownFields = new HashMap<>();
+
+  @JsonAnyGetter
+  public Map<String, Map<String, ZooKeeperStat>> unknownProperties() {
+    return unknownFields;
+  }
+
+  @JsonAnySetter
+  public void setUnknownProperty(String field, Map<String, ZooKeeperStat> value) {
+    unknownFields.put(field, value);
+  }
+}
